Fixes #129
